Unify the managed and object vtables
Allow for nullptr entries in the vtable. To nevertheless
get some decent error checking if one of the methods is
reimplemented, use a base class for Managed that contains
a full set of the vtable entries all being nullptr's.

+++ src/qml/jsruntime/qv4string.cpp
@@ -257,7 +257,7 @@ PropertyKey StringOrSymbol::toPropertyKey() const {
     return propertyKey();
 }
 
-uint String::getLength(const Managed *m)
+qint64 String::getLength(const Managed *m)
 {
     return static_cast<const String *>(m)->d()->length();
 }
